fix: fixes and improvments
