+++ contracts/croncat-agents/src/lib.rs
@@ -1,7 +1,7 @@
 #![doc = include_str!(concat!(env!("CARGO_MANIFEST_DIR"), "/README.md"))]
 
 pub mod contract;
-pub mod distro;
+pub mod distributor;
 pub mod error;
 mod external;
 pub mod msg;
